// src/app/main/models/Notice.ts
import Member from "./Member";
import { NoticeTemplate } from "./NoticeTemplate";

export interface Notice {
  id?: number; // Optional, auto-generated by the server
  notificationMethod: string; // Notification methods like SMS, Email, etc.
  receiver: string; // Receiver information (e.g., member identifier)
  subject: string; // Subject of the notice
  type: string; // Type of the notice (e.g., Scheduled Maintenance)
  importance: string; // Importance level (e.g., High, Medium, Low)
  noticeTemplate: NoticeTemplate; // Reference to the associated NoticeTemplate
  member: Member; // Reference to the associated Member
}